// CodeLabFrontend/src/app/shared/enums/forma-pagamento.enum.ts
export enum EFormaPagamento {
  DINHEIRO = 1,
  CARTAO_DEBITO = 2,
  CARTAO_CREDITO = 3,
  PIX = 4,
  BOLETO = 5,
  CHEQUE = 6,
  TRANSFERENCIA = 7,
  OUTROS = 8,
}

export enum EFormaPagamentoDescricao {
  DINHEIRO = 'Dinheiro',
  CARTAO_DEBITO = 'Cartão de Débito',
  CARTAO_CREDITO = 'Cartão de Crédito',
  PIX = 'PIX',
  BOLETO = 'Boleto',
  CHEQUE = 'Cheque',
  TRANSFERENCIA = 'Transferência',
  OUTROS = 'Outros',
}